third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-lifetime-mediakeysession-release-noreference.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeySession lifetime after release() without references</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <script>
            // MediaKeySessions remain as long as:
            //   JavaScript has a reference to it
            //   OR (MediaKeys is around
            //       AND the session has not received a close() event)
            async_test(function(test)
            {
                gc();
                var initDataType;
                var initData;
                var startingMediaKeysCount = window.internals.mediaKeysCount();
                var startingMediaKeySessionCount = window.internals.mediaKeySessionCount();

                function numMediaKeysCreated()
                {
                    return window.internals.mediaKeysCount() - startingMediaKeysCount;
                }

                function numMediaKeySessionCreated()
                {
                    return window.internals.mediaKeySessionCount() - startingMediaKeySessionCount;
                }

                // Create 2 sessions.
                var mediaKeys;
                var mediaKeySession1;
                var mediaKeySession2;

                navigator.requestMediaKeySystemAccess('org.w3.clearkey', getSimpleConfiguration()).then(function(access) {
                    initDataType = access.getConfiguration().initDataTypes[0];
                    initData = getInitData(initDataType);
                    return access.createMediaKeys();
                }).then(function(result) {
                    mediaKeys = result;

                    assert_equals(numMediaKeysCreated(), 1, 'MediaKeys.create()');
                    assert_equals(numMediaKeySessionCreated(), 0, 'MediaKeys.create()');

                    mediaKeySession1 = mediaKeys.createSession();
                    return mediaKeySession1.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession1.sessionId && mediaKeySession1.sessionId.length > 0);

                    assert_equals(numMediaKeysCreated(), 1, 'MediaKeys.createSession(1)');
                    assert_equals(numMediaKeySessionCreated(), 1, 'MediaKeys.createSession(1)');

                    mediaKeySession2 = mediaKeys.createSession();
                    return mediaKeySession2.generateRequest(initDataType, initData);
                }).then(function() {
                    assert_true(mediaKeySession2.sessionId && mediaKeySession2.sessionId.length > 0);

                    assert_equals(numMediaKeysCreated(), 1, 'mediaKeys.createSession(2)');
                    assert_equals(numMediaKeySessionCreated(), 2, 'mediaKeys.createSession(2)');
                }).then(function(result) {
                    // Run gc(). All sessions should remain as we have a
                    // reference to each one.
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 1, 'After gc()');
                    assert_equals(numMediaKeySessionCreated(), 2, 'After gc()');

                    // Close the sessions. Once the close() event is received,
                    // they should get garbage collected as there are no JS
                    // references to them.
                    var promise = mediaKeySession1.close();
                    mediaKeySession1 = null;
                    return promise;
                }).then(function(result) {
                    // Give time so that the close event can be processed by
                    // MediaKeySession.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 1, 'mediaKeySession1 not collected');
                    assert_equals(numMediaKeySessionCreated(), 1, 'mediaKeySession1 not collected');

                    var promise = mediaKeySession2.close();
                    mediaKeySession2 = null;
                    return promise;
                }).then(function(result) {
                    // Provide time for the mediaKeySession2 close event to be
                    // handled.
                    return delayToAllowEventProcessingPromise();
                }).then(function(result) {
                    return createGCPromise();
                }).then(function(result) {
                    assert_equals(numMediaKeysCreated(), 1, 'mediaKeySession2 not collected');
                    assert_equals(numMediaKeySessionCreated(), 0, 'mediaKeySession2 not collected');

                    assert_not_equals(mediaKeys, null);
                    test.done();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'MediaKeySession lifetime after release() without references');
        </script>
    </body>
</html>
